C3: move the sync observer into the data engine
data and sync were one subsystem in two directories: repo.ts imported
@/sync/observer while observer/* imported BlockCache, handleStore,
treeQueries, and the block schema back out of @/data — a 35-file SCC and
a directory layout that lied about the dependency direction.

Move src/sync/observer/* into src/data/internals/syncObserver/ — it is the
download half of the engine repo.ts already owns via startSyncObserver.
The sync->data import statements all lived in observer/*, so they collapse
into intra-data imports; the data->sync edge from repo.ts is gone too.

To keep src/sync genuinely lower (keys/crypto/transform are data-free), the
sync-seam vocabulary the §6 resolver shares with the observer
(Materializability / GetMaterializability) moves into @/sync/transform,
so sync/keys/resolver.ts no longer reaches up into the observer for a type.

Also build the §6 sync resolver once per user in repoProvider and expose
syncObserverDepsFor(userId): the upload connector and the Repo's download
observer now draw from the same resolver instead of context/repo.tsx
re-deriving a second one against the same key store + mode pins.

Closes #138
