+++ src/data/internals/syncObserver/observer.test.ts
@@ -11,8 +11,8 @@ import { BlockCache } from '@/data/blockCache'
 import { createTestDb, resetTestDb, type TestDb } from '@/data/test/createTestDb'
 import { startBlocksSyncedObserver, type BlocksSyncedObserver } from './observer.js'
 import type { GetMaterializability, Materializability } from './materialize.js'
-import { encodeForWire, type GetCek } from '../transform.js'
-import { generateWorkspaceKeyBytes, importWorkspaceKey } from '../crypto/workspaceKey.js'
+import { encodeForWire, type GetCek } from '@/sync/transform.js'
+import { generateWorkspaceKeyBytes, importWorkspaceKey } from '@/sync/crypto/workspaceKey.js'
 import type { ChangeNotification } from '@/data/internals/handleStore'
 import type { BlockData, CycleDetectedEvent } from '@/data/api'
 import { systemAuthor } from '@/data/api'
